[feature/DATEPICK-207]: Course api 마무리

+++ src/main/java/app/hdj/datepick/domain/course/service/CourseService.java
@@ -193,6 +193,10 @@ public class CourseService {
         }
 
         String imageUrl = course.getImageUrl();
+        if (imageUrl == null) {
+            throw new CustomException(ErrorCode.FILE_NOT_EXISTS);
+        }
+
         course.setImageUrl(null);
         fileService.remove(imageUrl);
     }
